updates to st play
