+++ t048_trainer.py
@@ -11,7 +11,7 @@ import tensorflow as tf
 
 from pynput import keyboard
 
-def dist_infer_action(sess,frames,ops,phs):
+def infer_action(sess,frames,ops_and_tens):
     """
         Function infers action from the inference network
     
@@ -24,24 +24,12 @@ def dist_infer_action(sess,frames,ops,phs):
             q: float. Q value of the action
     """
 
-    run_metadata = tf.RunMetadata()
-
-    action = ops['action']
-    x1 = phs['x1']
-    q = ops['q_vals_pr']
-
-    s_img1 = phs['s_img1']
-    s_a = phs['s_a']
-    s_r = phs['s_r']
-    s_img2 = phs['s_img2']
-    print("Before infer")
-    
-    #a,q = sess.run([action,q],{x1: [frames]})
-    #a,q = sess.run([action,q],{x1: [frames]},options=tf.RunOptions(trace_level=tf.RunOptions.FULL_TRACE),run_metadata=run_metadata)
-    a,q = sess.run([action,q],{s_img1: np.random.rand(5,100,100,4),s_a: np.random.rand(5,1).astype(np.float16), s_r: np.random.rand(5,1).astype(np.float16),s_img2: np.random.rand(5,100,100,4),x1: [frames]})
-    #a,q = sess.run([action],{x1: [np.zeros((100,100,4))]})
-    print("QQQ",q)
-    return a,q
+    s1 = ops_and_tens['s1']
+    s2 = ops_and_tens['s2']
+    r = ops_and_tens['r']
+    action = ops_and_tens['action']
+    a = sess.run([action],{s1: frames,s2: frames,r: [0]})
+    return a
 
 def send_action_to_game_controller(game,phi1,a,reward):
     """
@@ -219,34 +207,7 @@ def process_seq(seq):
     return np_f
 
 
-
-def dist_add_to_queue(sess,batch_size,ops,phs):
-    """
-        Function adds new training sequences to train queue
-        
-        args:
-            sess: Tensorflow Session.
-            batch_size: int. Training batch size.
-            ops:   Dict. Contains relevent tensorflow opeartions
-            phs: Dict. Contains relevent tensorflow placeholders
-        return:
-    
-    """
-    #Grab the enqueue operation and placeholders
-    enqueue_op = ops['enqueue_op']
-    s_img1 = phs['s_img1']
-    s_a = phs['s_a']
-    s_r = phs['s_r']
-    s_img2 = phs['s_img2']
-    x1 = phs['x1']
-
-    #Grab training batch
-    seq_n = random_minibatch_sample(batch_size)
-    #Add to training queue
-    sess.run([enqueue_op],{s_img1: seq_n[0],s_a: seq_n[1], s_r: seq_n[2],s_img2: seq_n[3],x1: np.random.rand(1,100,100,4)})
-
-
-def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops,phs,gsheets):
+def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops_and_tens,gsheets):
     """
         Function responsible for playing the game
 
@@ -286,7 +247,6 @@ def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops,phs,gsh
         while (process_frames < frame_limit):
             reward = 0
             wait_for(1)
-            
             frame,bval = get_frame(game)
             fff =[]
             seq = []
@@ -302,9 +262,7 @@ def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops,phs,gsh
                     #print(greed,"greedy: ")
                     a = np.asarray(np.random.randint(0,4))
                 else:
-                    #a,q = np.array(dist_infer_action(sess,phi1,ops,phs)).astype(np.float16)
-                    a,q = dist_infer_action(sess,phi1,ops,phs)
-                    #print(a,q)
+                    a = infer_action(sess,phi1,ops_and_tens)
                 frame,stop_play,r,reward = send_action_to_game_controller(game,phi1,a,reward)
                 seq.append(a)
                 seq.append(r)
@@ -316,7 +274,9 @@ def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops,phs,gsh
                     break
                 phi1 = phi2
                 if (len(exp) > batch_size):
-                    dist_add_to_queue(sess,batch_size,ops,phs)
+                    execute_train_operation(sess,batch_size,ops_and_tens)
+                    #dist_add_to_queue(sess,batch_size,ops,phs) 
+                    #EEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEE execute train
                 if force_kill:
                     game.stop_play = True
             wait_for(.3)
@@ -334,7 +294,7 @@ def frame_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops,phs,gsh
     return
 
 
-def dist_play(sess,game,M,ops,phs):
+def play(sess,game,M,ops_and_tens):
     for i in range(0,M):
         reward = 0
         wait_for(1)
@@ -343,8 +303,7 @@ def dist_play(sess,game,M,ops,phs):
             frames1,test = get_4_frames(game)
             if  test:
                 break
-            #a,q= np.asarray(dist_infer_action(sess,frames,ops,phs)).astype(np.float16)
-            a,q = dist_infer_action(sess,frames1,ops,phs)
+            a = infer_action(sess,frames1,ops_and_tens)
             print(a)
             frames2,test,r,reward = send_action_to_game_controller(game,frames1,a,reward)
             if test:
@@ -359,6 +318,93 @@ def save_seq_img(seq):
     for i in range(0,len(seq)):
         Image.fromarray(np.squeeze(seq[i])).save("imgs/test"+str(i)+".png")
 
+def iter_train_reward(sess,game,frame_limit,greed_frames,batch_size,ops_and_tens,gsheets):
+    #TODO: Function needs to be modifed to train for game iterations
+    """
+        Function responsible for playing the game
+
+        args:
+            sess:   Tensorflow session.
+            game:   GameController instance.
+            frame_limit:    int. Number of frames to train for
+            greed_frames:   int. Number of frames where greed is present.
+            batch_size: int. Size of training batch.
+            ops:    Dict. Relevent tensorflow operations.
+            phs:    Dict. Relevent tensorflow placeholders.
+
+        returns:
+
+    """
+
+
+    global process_frames
+    gp = 0
+    #cv2.imshow("frame",np.zeros(shape=(110,110,1)))
+    #cv2.waitKey(1)
+    #winname = "frame"
+    #cv2.namedWindow(winname) 
+    #cv2.moveWindow(winname, 2700,300)
+
+    force_kill = False
+
+    def stop_training(key):
+        print(key)
+        if key == keyboard.Key.esc:
+            print("Kill loop")
+            nonlocal force_kill
+            force_kill = True
+            return
+
+    with keyboard.Listener(on_press=stop_training) as listener:
+        while (process_frames < frame_limit):
+            reward = 0
+            wait_for(1)
+            frame,bval = get_frame(game)
+            fff =[]
+            seq = []
+            seq.append(frame)
+            fff.append(frame)
+            phi1 = process_seq(seq)        
+            while not game.stop_play:
+                greed = get_greed(greed_frames,process_frames)
+                #cv2.imshow(winname,frame)
+                #cv2.waitKey(1)
+                r_a = np.random.random_sample(1)
+                if (r_a <= greed):
+                    #print(greed,"greedy: ")
+                    a = np.asarray(np.random.randint(0,4))
+                else:
+                    a = infer_action(sess,phi1,ops_and_tens)
+                    #print(a,q)
+                frame,stop_play,r,reward = send_action_to_game_controller(game,phi1,a,reward)
+                seq.append(a)
+                seq.append(r)
+                seq.append(frame)
+                fff.append(frame)
+                phi2 = process_seq(seq)
+                store_exp((phi1,np.array(a).astype(np.uint8),np.array(r).astype(np.float16),phi2))
+                if stop_play:
+                    break
+                phi1 = phi2
+                if (len(exp) > batch_size):
+                    #dist_add_to_queue(sess,batch_size,ops,phs) 
+                    #EEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEE execute train
+                    x = 5
+                if force_kill:
+                    game.stop_play = True
+            wait_for(.3)
+            game.reward = 0
+            if force_kill:
+                break    
+            game.stop_play = False
+            gp = gp+1
+            if (gp % 50 == 0):
+                print("Exp size: ", len(exp))
+                print("Number process Frames: ",process_frames)
+                print("greed: ",greed)
+            save_seq_img(fff)
+        listener.stop()
+    return
